+++ kubernetes-model/kubernetes-model-common/src/main/java/io/fabric8/kubernetes/model/annotation/Namespaced.java
@@ -15,6 +15,10 @@
  */
 package io.fabric8.kubernetes.model.annotation;
 
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+
+@Retention(RetentionPolicy.RUNTIME)
 public @interface Namespaced {
   boolean value();
 }