+++ kubernetes-tests/src/test/java/io/fabric8/kubernetes/client/mock/DefaultSharedIndexInformerTest.java
@@ -16,6 +16,10 @@
 package io.fabric8.kubernetes.client.mock;
 
 import io.fabric8.kubernetes.api.model.ListMetaBuilder;
+import io.fabric8.kubernetes.api.model.Namespace;
+import io.fabric8.kubernetes.api.model.NamespaceBuilder;
+import io.fabric8.kubernetes.api.model.NamespaceList;
+import io.fabric8.kubernetes.api.model.NamespaceListBuilder;
 import io.fabric8.kubernetes.api.model.ObjectMetaBuilder;
 import io.fabric8.kubernetes.api.model.Pod;
 import io.fabric8.kubernetes.api.model.PodBuilder;
@@ -25,6 +29,14 @@ import io.fabric8.kubernetes.api.model.Status;
 import io.fabric8.kubernetes.api.model.StatusBuilder;
 import io.fabric8.kubernetes.api.model.WatchEvent;
 import io.fabric8.kubernetes.api.model.WatchEventBuilder;
+import io.fabric8.kubernetes.api.model.apps.Deployment;
+import io.fabric8.kubernetes.api.model.apps.DeploymentBuilder;
+import io.fabric8.kubernetes.api.model.apps.DeploymentList;
+import io.fabric8.kubernetes.api.model.apps.DeploymentListBuilder;
+import io.fabric8.kubernetes.api.model.rbac.ClusterRoleBinding;
+import io.fabric8.kubernetes.api.model.rbac.ClusterRoleBindingBuilder;
+import io.fabric8.kubernetes.api.model.rbac.ClusterRoleBindingList;
+import io.fabric8.kubernetes.api.model.rbac.ClusterRoleBindingListBuilder;
 import io.fabric8.kubernetes.client.KubernetesClient;
 import io.fabric8.kubernetes.client.dsl.base.CustomResourceDefinitionContext;
 import io.fabric8.kubernetes.client.dsl.base.OperationContext;
@@ -36,15 +48,18 @@ import io.fabric8.kubernetes.client.mock.crd.PodSetList;
 import io.fabric8.kubernetes.client.mock.crd.PodSetSpec;
 import io.fabric8.kubernetes.client.server.mock.KubernetesServer;
 import org.junit.Rule;
+import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.migrationsupport.rules.EnableRuleMigrationSupport;
 
 import java.net.HttpURLConnection;
 import java.util.Arrays;
-import java.util.concurrent.atomic.AtomicBoolean;
+import java.util.Collections;
+import java.util.concurrent.CountDownLatch;
+import java.util.concurrent.TimeUnit;
 
 import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertTrue;
+import static org.junit.Assert.assertFalse;
 
 @EnableRuleMigrationSupport
 public class DefaultSharedIndexInformerTest {
@@ -56,32 +71,35 @@ public class DefaultSharedIndexInformerTest {
       "401: The event in requested index is outdated and cleared (the requested history has been cleared [3/1]) [2]")
     .build();
   static final WatchEvent outdatedEvent = new WatchEventBuilder().withStatusObject(outdatedStatus).build();
+  static final Long WATCH_EVENT_EMIT_TIME = 50L;
+  static final Long OUTDATED_WATCH_EVENT_EMIT_TIME = 50L;
+  static final long RESYNC_PERIOD = 100L;
 
   @Test
   public void testNamespacedPodInformer() throws InterruptedException {
     String startResourceVersion = "1000", endResourceVersion = "1001";
 
     server.expect().withPath("/api/v1/namespaces/test/pods")
-      .andReturn(200, new PodListBuilder().withNewMetadata().withResourceVersion(startResourceVersion).endMetadata().withItems(Arrays.asList()).build()).once();
+      .andReturn(200, new PodListBuilder().withNewMetadata().withResourceVersion(startResourceVersion).endMetadata().withItems(Collections.emptyList()).build()).once();
     server.expect().withPath("/api/v1/namespaces/test/pods?resourceVersion=" + startResourceVersion + "&watch=true")
       .andUpgradeToWebSocket()
       .open()
-      .waitFor(1000)
+      .waitFor(WATCH_EVENT_EMIT_TIME)
       .andEmit(new WatchEvent(new PodBuilder().withNewMetadata().withName("pod1").withResourceVersion(endResourceVersion).endMetadata().build(), "ADDED"))
-      .waitFor(2000)
+      .waitFor(OUTDATED_WATCH_EVENT_EMIT_TIME)
       .andEmit(outdatedEvent).done().always();
 
     KubernetesClient client = server.getClient();
     SharedInformerFactory factory = client.informers();
-    SharedIndexInformer<Pod> podInformer = factory.sharedIndexInformerFor(Pod.class, PodList.class,4000L);
+    SharedIndexInformer<Pod> podInformer = factory.sharedIndexInformerFor(Pod.class, PodList.class,RESYNC_PERIOD);
 
-    AtomicBoolean foundExistingPod = new AtomicBoolean(false);
+    CountDownLatch foundExistingPod = new CountDownLatch(1);
     podInformer.addEventHandler(
       new ResourceEventHandler<Pod>() {
         @Override
         public void onAdd(Pod obj) {
           if (obj.getMetadata().getName().equalsIgnoreCase("pod1")) {
-            foundExistingPod.set(true);
+            foundExistingPod.countDown();
           }
         }
 
@@ -91,11 +109,10 @@ public class DefaultSharedIndexInformerTest {
         @Override
         public void onDelete(Pod oldObj, boolean deletedFinalStateUnknown) { }
       });
-    Thread.sleep(1000L);
     factory.startAllRegisteredInformers();
 
-    Thread.sleep(5000L);
-    assertEquals(true, foundExistingPod.get());
+    foundExistingPod.await(1, TimeUnit.SECONDS);
+    assertEquals(0L, foundExistingPod.getCount());
     assertEquals(endResourceVersion, podInformer.lastSyncResourceVersion());
 
     factory.stopAllRegisteredInformers();
@@ -110,22 +127,22 @@ public class DefaultSharedIndexInformerTest {
     server.expect().withPath("/api/v1/namespaces/test/pods?resourceVersion=" + startResourceVersion + "&watch=true")
       .andUpgradeToWebSocket()
       .open()
-      .waitFor(1000)
+      .waitFor(WATCH_EVENT_EMIT_TIME)
       .andEmit(new WatchEvent(new PodBuilder().withNewMetadata().withNamespace("test").withName("pod1").withResourceVersion(endResourceVersion).endMetadata().build(), "ADDED"))
-      .waitFor(2000)
+      .waitFor(OUTDATED_WATCH_EVENT_EMIT_TIME)
       .andEmit(outdatedEvent).done().always();
 
     KubernetesClient client = server.getClient();
     SharedInformerFactory factory = client.informers();
-    SharedIndexInformer<Pod> podInformer = factory.sharedIndexInformerFor(Pod.class, PodList.class, 4000L);
+    SharedIndexInformer<Pod> podInformer = factory.sharedIndexInformerFor(Pod.class, PodList.class, RESYNC_PERIOD);
 
-    AtomicBoolean foundExistingPod = new AtomicBoolean(false);
+    CountDownLatch foundExistingPod = new CountDownLatch(1);
     podInformer.addEventHandler(
       new ResourceEventHandler<Pod>() {
         @Override
         public void onAdd(Pod obj) {
           if (obj.getMetadata().getName().equalsIgnoreCase("pod1") && obj.getMetadata().getNamespace().equalsIgnoreCase("test")) {
-            foundExistingPod.set(true);
+            foundExistingPod.countDown();
           }
         }
 
@@ -135,11 +152,10 @@ public class DefaultSharedIndexInformerTest {
         @Override
         public void onDelete(Pod oldObj, boolean deletedFinalStateUnknown) { }
       });
-    Thread.sleep(1000L);
     factory.startAllRegisteredInformers();
 
-    Thread.sleep(5000L);
-    assertEquals(true, foundExistingPod.get());
+    foundExistingPod.await(1, TimeUnit.SECONDS);
+    assertEquals(0, foundExistingPod.getCount());
     assertEquals(endResourceVersion, podInformer.lastSyncResourceVersion());
 
     factory.stopAllRegisteredInformers();
@@ -150,30 +166,30 @@ public class DefaultSharedIndexInformerTest {
     String startResourceVersion = "1000", midResourceVersion = "1001", mid2ResourceVersion = "1002", endResourceVersion = "1003";
 
     server.expect().withPath("/api/v1/namespaces/test/pods")
-      .andReturn(200, new PodListBuilder().withNewMetadata().withResourceVersion(startResourceVersion).endMetadata().withItems(Arrays.asList()).build()).once();
+      .andReturn(200, new PodListBuilder().withNewMetadata().withResourceVersion(startResourceVersion).endMetadata().withItems(Collections.emptyList()).build()).once();
     server.expect().withPath("/api/v1/namespaces/test/pods?resourceVersion=" + startResourceVersion + "&watch=true")
       .andUpgradeToWebSocket()
       .open()
-      .waitFor(1000)
+      .waitFor(WATCH_EVENT_EMIT_TIME)
       .andEmit(new WatchEvent(new PodBuilder().withNewMetadata().withNamespace("test").withName("pod1").withResourceVersion(midResourceVersion).endMetadata().build(), "ADDED"))
-      .waitFor(2000)
+      .waitFor(OUTDATED_WATCH_EVENT_EMIT_TIME)
       .andEmit(outdatedEvent)
       .done().always();
     server.expect().withPath("/api/v1/namespaces/test/pods")
-      .andReturn(200, new PodListBuilder().withNewMetadata().withResourceVersion(mid2ResourceVersion).endMetadata().withItems(Arrays.asList()).build()).times(2);
+      .andReturn(200, new PodListBuilder().withNewMetadata().withResourceVersion(mid2ResourceVersion).endMetadata().withItems(Collections.emptyList()).build()).times(2);
     server.expect().withPath("/api/v1/namespaces/test/pods?resourceVersion=" + mid2ResourceVersion + "&watch=true")
       .andUpgradeToWebSocket()
       .open()
-      .waitFor(1000)
+      .waitFor(WATCH_EVENT_EMIT_TIME)
       .andEmit(new WatchEvent(new PodBuilder().withNewMetadata().withNamespace("test").withName("pod1").withResourceVersion(endResourceVersion).endMetadata().build(), "MODIFIED"))
       .done()
       .always();
 
     KubernetesClient client = server.getClient();
     SharedInformerFactory factory = client.informers();
-    SharedIndexInformer<Pod> podInformer = factory.sharedIndexInformerFor(Pod.class, PodList.class, 10000L);
+    SharedIndexInformer<Pod> podInformer = factory.sharedIndexInformerFor(Pod.class, PodList.class, RESYNC_PERIOD);
 
-    AtomicBoolean relistSuccessful = new AtomicBoolean(false);
+    CountDownLatch relistSuccessful = new CountDownLatch(1);
     podInformer.addEventHandler(
       new ResourceEventHandler<Pod>() {
         @Override
@@ -183,18 +199,17 @@ public class DefaultSharedIndexInformerTest {
         public void onUpdate(Pod oldObj, Pod newObj) {
           if (newObj.getMetadata().getName().equalsIgnoreCase("pod1") &&
             newObj.getMetadata().getResourceVersion().equalsIgnoreCase(endResourceVersion)) {
-            relistSuccessful.set(true);
+            relistSuccessful.countDown();
           }
         }
 
         @Override
         public void onDelete(Pod oldObj, boolean deletedFinalStateUnknown) { }
       });
-    Thread.sleep(1000L);
     factory.startAllRegisteredInformers();
 
-    Thread.sleep(10000L);
-    assertEquals(true, relistSuccessful.get());
+    relistSuccessful.await(7, TimeUnit.SECONDS);
+    assertEquals(0, relistSuccessful.getCount());
     assertEquals(endResourceVersion, podInformer.lastSyncResourceVersion());
 
     factory.stopAllRegisteredInformers();
@@ -208,14 +223,14 @@ public class DefaultSharedIndexInformerTest {
     server.expect().withPath("/api/v1/namespaces/test/pods?resourceVersion=" + startResourceVersion + "&watch=true")
       .andUpgradeToWebSocket()
       .open()
-      .waitFor(1000)
+      .waitFor(WATCH_EVENT_EMIT_TIME)
       .andEmit(new WatchEvent(new PodBuilder().withNewMetadata().withNamespace("test").withName("pod1").withResourceVersion(endResourceVersion).endMetadata().build(), "ADDED"))
-      .waitFor(2000)
+      .waitFor(OUTDATED_WATCH_EVENT_EMIT_TIME)
       .andEmit(outdatedEvent).done().once();
 
     KubernetesClient client = server.getClient();
     SharedInformerFactory factory = client.informers();
-    SharedIndexInformer<Pod> podInformer = factory.sharedIndexInformerFor(Pod.class, PodList.class, 4000L);
+    SharedIndexInformer<Pod> podInformer = factory.sharedIndexInformerFor(Pod.class, PodList.class, 2000L);
 
     podInformer.addEventHandler(
       new ResourceEventHandler<Pod>() {
@@ -228,33 +243,21 @@ public class DefaultSharedIndexInformerTest {
         @Override
         public void onDelete(Pod oldObj, boolean deletedFinalStateUnknown) { }
       });
-    Thread.sleep(1000L);
     factory.startAllRegisteredInformers();
 
     // Wait for resync period of pass
-    Thread.sleep(8000L);
-    assertEquals(false, podInformer.hasSynced());
+    Thread.sleep(RESYNC_PERIOD);
+    assertFalse(podInformer.hasSynced());
 
     factory.stopAllRegisteredInformers();
   }
 
   @Test
   public void testEventListeners() throws InterruptedException {
-    String startResourceVersion = "1000", endResourceVersion = "1001";
-    server.expect().withPath("/api/v1/namespaces/test/pods")
-      .andReturn(200, new PodListBuilder().withNewMetadata().withResourceVersion(startResourceVersion).endMetadata().withItems(Arrays.asList()).build()).once();
-    server.expect().withPath("/api/v1/namespaces/test/pods?resourceVersion=" + startResourceVersion + "&watch=true")
-      .andUpgradeToWebSocket()
-      .open()
-      .waitFor(1000)
-      .andEmit(new WatchEvent(new PodBuilder().withNewMetadata().withNamespace("test").withName("pod1").withResourceVersion(endResourceVersion).endMetadata().build(), "ADDED"))
-      .waitFor(2000)
-      .andEmit(outdatedEvent).done().once();
-
     KubernetesClient client = server.getClient();
     SharedInformerFactory factory = client.informers();
-    SharedIndexInformer<Pod> podInformer = factory.sharedIndexInformerFor(Pod.class, PodList.class, 4000L);
-    AtomicBoolean failureCallbackReceived = new AtomicBoolean(false);
+    SharedIndexInformer<Pod> podInformer = factory.sharedIndexInformerFor(Pod.class, PodList.class, 1000L);
+    CountDownLatch failureCallbackReceived = new CountDownLatch(1);
 
     podInformer.addEventHandler(
       new ResourceEventHandler<Pod>() {
@@ -268,13 +271,153 @@ public class DefaultSharedIndexInformerTest {
         public void onDelete(Pod oldObj, boolean deletedFinalStateUnknown) { }
       });
 
-    factory.addSharedInformerEventListener(exception -> failureCallbackReceived.set(true));
-    Thread.sleep(1000L);
+    factory.addSharedInformerEventListener(exception -> failureCallbackReceived.countDown());
     factory.startAllRegisteredInformers();
 
     // Wait for resync period of pass
-    Thread.sleep(8000L);
-    assertEquals(false, failureCallbackReceived.get());
+    failureCallbackReceived.await(1, TimeUnit.SECONDS);
+    assertEquals(0, failureCallbackReceived.getCount());
+
+    factory.stopAllRegisteredInformers();
+  }
+
+  @Test
+  @DisplayName("Should create Informer for Namespace resource")
+  public void testWithNamespacedResource() throws InterruptedException {
+    String startResourceVersion = "1000", endResourceVersion = "1001";
+
+    server.expect().withPath("/api/v1/namespaces")
+      .andReturn(200, new NamespaceListBuilder().withNewMetadata().withResourceVersion(startResourceVersion).endMetadata().withItems(Arrays.asList()).build()).once();
+    server.expect().withPath("/api/v1/namespaces?resourceVersion=" + startResourceVersion + "&watch=true")
+      .andUpgradeToWebSocket()
+      .open()
+      .waitFor(WATCH_EVENT_EMIT_TIME)
+      .andEmit(new WatchEvent(new NamespaceBuilder().withNewMetadata().withName("ns1").withResourceVersion(endResourceVersion).endMetadata().build(), "ADDED"))
+      .waitFor(OUTDATED_WATCH_EVENT_EMIT_TIME)
+      .andEmit(outdatedEvent).done().always();
+
+    KubernetesClient client = server.getClient();
+    SharedInformerFactory factory = client.informers();
+    SharedIndexInformer<Namespace> namespaceSharedIndexInformer = factory.sharedIndexInformerFor(
+      Namespace.class, NamespaceList.class,
+      // Provided Namespace to mimic Openshift scenario where context saves namespace
+      new OperationContext().withNamespace("myproject"),
+      1000L);
+
+    CountDownLatch foundExistingNamespace = new CountDownLatch(1);
+    namespaceSharedIndexInformer.addEventHandler(
+      new ResourceEventHandler<Namespace>() {
+        @Override
+        public void onAdd(Namespace obj) {
+          if (obj.getMetadata().getName().equalsIgnoreCase("ns1")) {
+            foundExistingNamespace.countDown();
+          }
+        }
+
+        @Override
+        public void onUpdate(Namespace oldObj, Namespace newObj) { }
+
+        @Override
+        public void onDelete(Namespace oldObj, boolean deletedFinalStateUnknown) { }
+      });
+    factory.startAllRegisteredInformers();
+
+    foundExistingNamespace.await(1, TimeUnit.SECONDS);
+    assertEquals(0, foundExistingNamespace.getCount());
+    assertEquals(endResourceVersion, namespaceSharedIndexInformer.lastSyncResourceVersion());
+
+    factory.stopAllRegisteredInformers();
+  }
+
+  @Test
+  @DisplayName("Should create informer for ClusterRoleBinding resource")
+  public void testWithClusterBindingInformer() throws InterruptedException {
+    String startResourceVersion = "1000", endResourceVersion = "1001";
+
+    server.expect().withPath("/apis/rbac.authorization.k8s.io/v1beta1/clusterrolebindings")
+      .andReturn(200, new ClusterRoleBindingListBuilder().withMetadata(new ListMetaBuilder().withResourceVersion(startResourceVersion).build()).withItems(Arrays.asList()).build()).once();
+    server.expect().withPath("/apis/rbac.authorization.k8s.io/v1beta1/clusterrolebindings?resourceVersion=" + startResourceVersion + "&watch=true")
+      .andUpgradeToWebSocket()
+      .open()
+      .waitFor(WATCH_EVENT_EMIT_TIME)
+      .andEmit(new WatchEvent(new ClusterRoleBindingBuilder().withNewMetadata().withName("crb1").withResourceVersion(endResourceVersion).endMetadata().build(), "ADDED"))
+      .waitFor(OUTDATED_WATCH_EVENT_EMIT_TIME)
+      .andEmit(outdatedEvent).done().always();
+
+    KubernetesClient client = server.getClient();
+    SharedInformerFactory factory = client.informers();
+    SharedIndexInformer<ClusterRoleBinding> clusterRoleBindingSharedIndexInformer = factory.sharedIndexInformerFor(
+      ClusterRoleBinding.class, ClusterRoleBindingList.class,
+      RESYNC_PERIOD);
+
+    CountDownLatch foundExistingClusterRoleBinding = new CountDownLatch(1);
+    clusterRoleBindingSharedIndexInformer.addEventHandler(
+      new ResourceEventHandler<ClusterRoleBinding>() {
+        @Override
+        public void onAdd(ClusterRoleBinding obj) {
+          if (obj.getMetadata().getName().equalsIgnoreCase("crb1")) {
+            foundExistingClusterRoleBinding.countDown();
+          }
+        }
+
+        @Override
+        public void onUpdate(ClusterRoleBinding oldObj, ClusterRoleBinding newObj) { }
+
+        @Override
+        public void onDelete(ClusterRoleBinding oldObj, boolean deletedFinalStateUnknown) { }
+      });
+    factory.startAllRegisteredInformers();
+
+    foundExistingClusterRoleBinding.await(1, TimeUnit.SECONDS);
+    assertEquals(0, foundExistingClusterRoleBinding.getCount());
+    assertEquals(endResourceVersion, clusterRoleBindingSharedIndexInformer.lastSyncResourceVersion());
+
+    factory.stopAllRegisteredInformers();
+  }
+
+  @Test
+  @DisplayName("Should create informer for Deployment resource")
+  public void testWithDeploymentInformer() throws InterruptedException {
+    String startResourceVersion = "1000", endResourceVersion = "1001";
+
+    server.expect().withPath("/apis/apps/v1/namespaces/ns1/deployments")
+      .andReturn(200, new DeploymentListBuilder().withMetadata(new ListMetaBuilder().withResourceVersion(startResourceVersion).build()).withItems(Arrays.asList()).build()).once();
+    server.expect().withPath("/apis/apps/v1/namespaces/ns1/deployments?resourceVersion=" + startResourceVersion + "&watch=true")
+      .andUpgradeToWebSocket()
+      .open()
+      .waitFor(WATCH_EVENT_EMIT_TIME)
+      .andEmit(new WatchEvent(new DeploymentBuilder().withNewMetadata().withName("deployment1").withResourceVersion(endResourceVersion).endMetadata().build(), "ADDED"))
+      .waitFor(OUTDATED_WATCH_EVENT_EMIT_TIME)
+      .andEmit(outdatedEvent).done().always();
+
+    KubernetesClient client = server.getClient();
+    SharedInformerFactory factory = client.informers();
+    SharedIndexInformer<Deployment> deploymentSharedIndexInformer = factory.sharedIndexInformerFor(
+      Deployment.class, DeploymentList.class,
+      new OperationContext().withNamespace("ns1"),
+      RESYNC_PERIOD);
+
+    CountDownLatch foundExistingDeployment = new CountDownLatch(1);
+    deploymentSharedIndexInformer.addEventHandler(
+      new ResourceEventHandler<Deployment>() {
+        @Override
+        public void onAdd(Deployment obj) {
+          if (obj.getMetadata().getName().equalsIgnoreCase("deployment1")) {
+            foundExistingDeployment.countDown();
+          }
+        }
+
+        @Override
+        public void onUpdate(Deployment oldObj, Deployment newObj) { }
+
+        @Override
+        public void onDelete(Deployment oldObj, boolean deletedFinalStateUnknown) { }
+      });
+    factory.startAllRegisteredInformers();
+
+    foundExistingDeployment.await(1, TimeUnit.SECONDS);
+    assertEquals(0, foundExistingDeployment.getCount());
+    assertEquals(endResourceVersion, deploymentSharedIndexInformer.lastSyncResourceVersion());
 
     factory.stopAllRegisteredInformers();
   }
@@ -288,9 +431,9 @@ public class DefaultSharedIndexInformerTest {
     server.expect().withPath("/api/v1/namespaces/ns1/pods?resourceVersion=" + startResourceVersion + "&watch=true")
       .andUpgradeToWebSocket()
       .open()
-      .waitFor(1000)
+      .waitFor(WATCH_EVENT_EMIT_TIME)
       .andEmit(new WatchEvent(new PodBuilder().withNewMetadata().withName("pod1").withResourceVersion(endResourceVersion).endMetadata().build(), "ADDED"))
-      .waitFor(2000)
+      .waitFor(OUTDATED_WATCH_EVENT_EMIT_TIME)
       .andEmit(outdatedEvent).done().always();
 
     KubernetesClient client = server.getClient();
@@ -298,15 +441,15 @@ public class DefaultSharedIndexInformerTest {
     SharedIndexInformer<Pod> podInformer = factory.sharedIndexInformerFor(
       Pod.class, PodList.class,
       new OperationContext().withNamespace("ns1"),
-      4000L);
+      100L);
 
-    AtomicBoolean foundExistingPod = new AtomicBoolean(false);
+    CountDownLatch foundExistingPod = new CountDownLatch(1);
     podInformer.addEventHandler(
       new ResourceEventHandler<Pod>() {
         @Override
         public void onAdd(Pod obj) {
           if (obj.getMetadata().getName().equalsIgnoreCase("pod1")) {
-            foundExistingPod.set(true);
+            foundExistingPod.countDown();
           }
         }
 
@@ -316,11 +459,10 @@ public class DefaultSharedIndexInformerTest {
         @Override
         public void onDelete(Pod oldObj, boolean deletedFinalStateUnknown) { }
       });
-    Thread.sleep(1000L);
     factory.startAllRegisteredInformers();
 
-    Thread.sleep(5000L);
-    assertTrue(foundExistingPod.get());
+    foundExistingPod.await(1, TimeUnit.SECONDS);
+    assertEquals(0, foundExistingPod.getCount());
     assertEquals(endResourceVersion, podInformer.lastSyncResourceVersion());
 
     factory.stopAllRegisteredInformers();
@@ -338,9 +480,9 @@ public class DefaultSharedIndexInformerTest {
     server.expect().withPath("/apis/demo.k8s.io/v1alpha1/namespaces/ns1/podsets?resourceVersion=" + startResourceVersion + "&watch=true")
       .andUpgradeToWebSocket()
       .open()
-      .waitFor(1000)
+      .waitFor(WATCH_EVENT_EMIT_TIME)
       .andEmit(new WatchEvent(getPodSet("podset1", endResourceVersion), "ADDED"))
-      .waitFor(2000)
+      .waitFor(OUTDATED_WATCH_EVENT_EMIT_TIME)
       .andEmit(outdatedEvent).done().always();
     KubernetesClient client = server.getClient();
 
@@ -352,14 +494,14 @@ public class DefaultSharedIndexInformerTest {
       .build();
 
     SharedInformerFactory sharedInformerFactory = client.informers();
-    SharedIndexInformer<PodSet> podInformer = sharedInformerFactory.sharedIndexInformerForCustomResource(crdContext, PodSet.class, PodSetList.class, new OperationContext().withNamespace("ns1"), 1 * 60 * 1000);
-    AtomicBoolean foundExistingPodSet = new AtomicBoolean(false);
+    SharedIndexInformer<PodSet> podInformer = sharedInformerFactory.sharedIndexInformerForCustomResource(crdContext, PodSet.class, PodSetList.class, new OperationContext().withNamespace("ns1"), 1 * 60 * WATCH_EVENT_EMIT_TIME);
+    CountDownLatch foundExistingPodSet = new CountDownLatch(1);
     podInformer.addEventHandler(
       new ResourceEventHandler<PodSet>() {
         @Override
         public void onAdd(PodSet podSet) {
           if (podSet.getMetadata().getName().equalsIgnoreCase("podset1")) {
-            foundExistingPodSet.set(true);
+            foundExistingPodSet.countDown();
           }
         }
 
@@ -369,11 +511,10 @@ public class DefaultSharedIndexInformerTest {
         @Override
         public void onDelete(PodSet podSet, boolean deletedFinalStateUnknown) { }
       });
-    Thread.sleep(1000L);
     sharedInformerFactory.startAllRegisteredInformers();
 
-    Thread.sleep(5000L);
-    assertTrue(foundExistingPodSet.get());
+    foundExistingPodSet.await(1, TimeUnit.SECONDS);
+    assertEquals(0, foundExistingPodSet.getCount());
     assertEquals(endResourceVersion, podInformer.lastSyncResourceVersion());
 
     sharedInformerFactory.stopAllRegisteredInformers();