+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### 4.10-SNAPSHOT
 #### Bugs
 * Fix #2205: Event model classes from core v1 have been lost
+* Fix #2226: SharedIndexInformer for non-namespaced resources not working
 * Fix #2201: Uberjar doesn't contain model classes anymore
 * Fix #2066: Uber Jar includes merged service entry for multiple implementations of the same interface
 * Fix #2195: Annotation processors and build time dependencies transitive